backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
python-multipart==0.0.20
pycryptodome==3.23.0